event data updated

+++ merchant-portal-lagom-app-api/src/main/scala/com/squareoneinsights/merchantportallagomapp/api/response/MerchantTransactionResp.scala
@@ -7,7 +7,7 @@ import java.sql.Timestamp
 case class MerchantTransactionResp(txnId:String,
                                    caseRefNo:String,
                                    txnTimestamp:String,
-                                   txnAmount:Int,
+                                   txnAmount:Double,
                                    ifrmVerdict:String,
                                    investigationStatus:String,
                                    channel:String,
