+++ merchant-portal-lagom-app-impl/src/main/scala/com/squareoneinsights/merchantportallagomapp/impl/model/MerchantTransactions.scala
@@ -13,7 +13,7 @@ case class MerchantTransaction(
     txnId: String,
     caseRefNo: String,
     txnTimestamp: Timestamp,
-    txnAmount: Int,
+    txnAmount: Double,
     ifrmVerdict: String,
     investigationStatus: String,
     channel: String,
